Protect bamboo during hudson build
Hudson does not have scope of other projects when running the build phase. As a result the unit tests fail when you trry to execute sharedlib code from tviplayer. This change protects bamboo from this by checking if the tviplayer class exists before executing it. This is good practice regardless of the hudson process as it protects us from changes in tviplayer.