Fix for Zend_Wildfire issues
This is one of those bugs with a tiny non-fatal coding error that causes a horrible chain reaction that falls through a dozen files and is very tricky to debug.

Corrects error
```
Notice 2048 - Declaration of BBC_Service_BambooFake::fetch() should be
             compatible with that of BBC_Service_Bamboo::fetch()
             /mnt/hgfs/workspace/bamboo/src/main/lib/BBC/Service/BambooFake.php (72)
```

The impact of this bug is to fix the Zend wildfire bugs that have been crippling the test environment.

By tracing the wildfire errors you see that the chain exits the tviplayer codebase at `BBC_Tviplayer_DataSources_KvStoreBase::getKvStoreClient() /data/webapps/bbc.co.uk/tviplayer/webapp/php/lib/BBC/Tviplayer/DataSources/KvStoreBase.php:42`. After this step it enters the PAL framework and starts the process of loading services via the service broker. The fact that KvStoreBase is triggering this bug is misleading - it just happens to be the first point in code that executes the service broker.

During the service broker contruction all include paths are checked for possible services. It does this by parsing all the filenames in the BBC/Service directory and checking if a class exists which matches that file. This fires up the php autoloader. The autoloader finds BambooFake (which we dont even use on the test environment) and triggers the declaration notice above. Normally this wouldnt cause a fatal error.

However on test we have some extra logging for notices (https://github.com/iplayer/responsive-web/blob/develop/webapp/php/lib/BBC/Tviplayer/ErrorHandler.php?source=cc#L24). This errorhandler gets executed as the process wide errorhandler when the notice is triggered - as a result a fatal is thrown because `BBC_Tviplayer_Log` doesnt exist as far as Bamboo is concerned - its not in sharedlib.