Add get{size}Guidance getters for Versions
Instead of IBL returning an object with a single 'text' element, they now correctly return a small/medium/large (synopsis style) object.
This change exposes those values in the same way as the synopsis.

Note: This is a breaking change for anything using getGuidance(), and as such will be followed up with a responsive-web pull request.